requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.10
pillow==10.2.0
pycryptodome==3.20.0
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing-extensions==4.9.0
